cafe-api/src/test/java/ca/ulaval/glo4002/cafe/small/cafe/api/layout/assembler/LayoutResponseAssemblerTest.java
package ca.ulaval.glo4002.cafe.small.cafe.api.layout.assembler;

import org.junit.jupiter.api.Test;

import ca.ulaval.glo4002.cafe.api.layout.assembler.LayoutResponseAssembler;
import ca.ulaval.glo4002.cafe.api.layout.response.LayoutResponse;
import ca.ulaval.glo4002.cafe.application.layout.payload.LayoutPayload;
import ca.ulaval.glo4002.cafe.domain.Cafe;
import ca.ulaval.glo4002.cafe.fixture.CafeFixture;
import ca.ulaval.glo4002.cafe.fixture.CubeResponseFixture;
import ca.ulaval.glo4002.cafe.fixture.LayoutResponseFixture;

import static org.junit.jupiter.api.Assertions.assertEquals;

public class LayoutResponseAssemblerTest {
    private static final Cafe A_CAFE = new CafeFixture().build();

    private final LayoutResponseAssembler layoutResponseAssembler = new LayoutResponseAssembler();

    @Test
    public void givenACafe_whenAssemblingLayoutResponse_shouldReturnValidLayoutResponse() {
        LayoutPayload layoutPayload = LayoutPayload.fromCafe(A_CAFE);
        LayoutResponse expectedLayoutResponse = new LayoutResponseFixture().withName(A_CAFE.getName().value())
            .withCubeResponses(new CubeResponseFixture().createCubeResponsesWithCubes(A_CAFE.getLayout().getCubes())).build();

        LayoutResponse actualLayoutResponse = layoutResponseAssembler.toLayoutResponse(layoutPayload);

        assertEquals(expectedLayoutResponse.name(), actualLayoutResponse.name());
        assertEquals(expectedLayoutResponse.cubes(), actualLayoutResponse.cubes());
    }
}
